chore(fsrs): move legacy scheduler
