+++ packages/fsrs/bench/fuzzing.bench.ts
@@ -6,7 +6,7 @@ import {
 import { dateChrono } from '@open-spaced-repetition/srs-kit/chrono/date'
 import { alea } from 'ts-fsrs/alea'
 import { createEmptyCard } from 'ts-fsrs/default'
-import { fsrs } from 'ts-fsrs/fsrs'
+import { fsrs } from 'ts-fsrs/legacy/fsrs'
 import {
   type FuzzingRng,
   fnv1aMulberry32Rng,
