+++ packages/fsrs/bench/learning-steps.bench.ts
@@ -5,7 +5,7 @@ import {
 } from '@open-spaced-repetition/srs-kit'
 import { dateChrono } from '@open-spaced-repetition/srs-kit/chrono/date'
 import { createEmptyCard } from 'ts-fsrs/default'
-import { fsrs } from 'ts-fsrs/fsrs'
+import { fsrs } from 'ts-fsrs/legacy/fsrs'
 import {
   ConvertStepUnitToMinutes,
   calculateLearningSteps,
