--- packages/fsrs/jsr.json
@@ -1,18 +1,0 @@
-{
-  "$schema": "https://jsr.io/schema/config-file.v1.json",
-  "name": "@fsrs/ts-fsrs",
-  "version": "5.2.3",
-  "license": "MIT",
-  "exports": {
-    ".": "./src/index.ts"
-  },
-  "publish": {
-    "include": [
-      "src/**/*.ts"
-    ],
-    "exclude": [
-      "__tests__"
-    ]
-  }
-}
-
