+++ packages/fsrs/src/legacy/__tests__/fixed/calc-elapsed-days.test.ts
@@ -61,14 +61,15 @@ test('SSE use next_state', () => {
     const current = new Date(rids[i])
     const rating = ratings[i]
     const delta_t = dateDiffInDays(last, current)
+    if (rating === Rating.Manual) {
+      continue
+    }
     const nextStates = f.model.step({
       memoryState,
       rating,
       elapsedDays: delta_t,
     })
-    if (rating !== 0) {
-      last = new Date(rids[i])
-    }
+    last = new Date(rids[i])
 
     console.debug(
       rids[i + 1],
