+++ packages/fsrs/src/legacy/__tests__/middlewares/learning-steps/regression.spec.ts
@@ -3,7 +3,7 @@
 import { Rating, State } from '@open-spaced-repetition/srs-kit'
 import { describe, expect, it } from 'vitest'
 import { createEmptyCard } from '@/default.js'
-import { fsrs } from '@/fsrs.js'
+import { fsrs } from '@/legacy/fsrs.js'
 
 describe('learning-step regression #311', () => {
   const scheduler = fsrs({
