+++ packages/fsrs/src/legacy/impl/long_term_scheduler.ts
@@ -1,16 +1,16 @@
-import { AbstractScheduler } from '../abstract_scheduler.js'
-import { TypeConvert } from '../convert.js'
-import { date_scheduler } from '../help.js'
-import { withFuzzing } from '../middlewares/fuzzing/core.js'
-import { calculateScheduleDays } from '../middlewares/monotonic-interval/core.js'
+import { TypeConvert } from '@/convert.js'
+import { date_scheduler } from '@/help.js'
+import { withFuzzing } from '@/middlewares/fuzzing/core.js'
+import { calculateScheduleDays } from '@/middlewares/monotonic-interval/core.js'
 import {
   type Card,
   type Grade,
   Rating,
   type RecordLogItem,
   State,
-} from '../models.js'
-import type { int } from '../types.js'
+} from '@/models.js'
+import type { int } from '@/types.js'
+import { AbstractScheduler } from '../abstract_scheduler.js'
 
 export default class LongTermScheduler extends AbstractScheduler {
   protected override newState(grade: Grade): RecordLogItem {
