+++ packages/fsrs/tsconfig.json
@@ -4,8 +4,11 @@
     "outDir": "./dist",
     "rootDir": "./src",
     "paths": {
-      "@/*": ["./src/*"]
-    }
+      "@/*": ["./src/*"],
+      "ts-fsrs": ["./src/index.ts"],
+      "ts-fsrs/*": ["./src/*"]
+    },
+    "types": ["vitest/globals", "node"]
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist", "__tests__"]
